Install drainer through NSSM (#20)
This PR installs the drainer through NSSM. Installing through powershell cmdlets or using sc.exe proved unnecessarily fiddly and lacked configuration options such as setting the working directory for the service.

Using NSSM reduces these problems as it can just run a script.

There is a risk the choco install might fail but this should be minimal and if it does fail intermittently it is not a huge problem as this is not a on the critical path. Just provides a tidying up service!