Add Terraform for AIB without a PIP (#22)
We use the Azure Image Builder Service(AIB) to generate the Linux Image. The configuration for the image lives at https://github.com/UKHO/vm-agent-aib

Previously the image has been built by the AIB creating an RG and VM inside the AzDo Live subscription and attaching a PIP to it and running the scripts found in the vm-agent-aib repo. As we cannot create/assign PIPs anymore this process does not work and we cannot add an exclusion to allow a PIP as the RG is generated with a random name each time.

Instead we want to use the new functionality of AIB to use an existing VNet, AIB will setup a LB, proxy VM and private link to be able to create the image without a PIP.

Co-authored-by: abrandon <abrandon@ukho.gov.uk>
Co-authored-by: Marc Denman <MarcDenman@users.noreply.github.com>