Re-add Python (#23)
Originally the python script caused the build to hang so needed to be removed. This PR adds it back in as after investigation it appears to be related to the Ubuntu base image.

After checking the logs, I have seen it is hanging with Checking init scripts which lead to this SO question - https://stackoverflow.com/questions/58219984/ec2-user-data-hanging-on-checking-init-scripts-ubuntu-php

This PR adds in the python script again and moves from a pinned version to the latest version fixes the problem and ensures we are more up to date.